:computer: Implements a macro to inject photo form helpers
